[Refactor] Write Command (Scrum 수정) - InvalidArgumentException 추가 - Test 코드 Param 개수만 유효성 체크
