+++ tests/test_write.py
@@ -1,0 +1,47 @@
+import pytest
+
+from src.commands.command_action import InvalidArgumentException
+from src.commands.write import WriteCommand
+from src.ssd import VirtualSSD
+
+
+@pytest.fixture
+def mock_ssd(mocker):
+    mock_ssd = mocker.Mock(spec=VirtualSSD)
+    return mock_ssd
+
+
+@pytest.mark.parametrize('test_address', [0, 15, 99])
+def test_write_command_성공(test_address, mock_ssd):
+    test_value = "0x12345678"
+
+    write_cmd = WriteCommand(mock_ssd, test_address, test_value)
+    write_cmd.run()
+
+    mock_ssd.write.assert_called_once()
+
+
+@pytest.mark.parametrize('test_value, error_param', [('0x1234567Z', -1), ('0x123456Z8', 'c'),
+                                                     ('0x12345Z78', 1), ('0x1234Z678', 100),
+                                                     ('0x123Z5678', 0.1), ('0x12Z45678', 123),
+                                                     ('0x1Z345678', 1), ('0xZ2345678', 0)])
+def test_write_command_유효성검사_Param개수_초과(test_value, error_param, mock_ssd):
+    test_address = 0
+
+    write_cmd = WriteCommand(mock_ssd, test_address, test_value, error_param)
+
+    with pytest.raises(InvalidArgumentException):
+        write_cmd.run()
+
+    assert not write_cmd.validate()
+    mock_ssd.write.assert_not_called()
+
+
+def test_write_command_유효성검사_Param개수_부족(mock_ssd):
+    write_cmd = WriteCommand(mock_ssd)
+
+    with pytest.raises(InvalidArgumentException):
+        write_cmd.run()
+
+    assert not write_cmd.validate()
+    mock_ssd.write.assert_not_called()
